security/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0  # JWT authentication
passlib[bcrypt]>=1.7.4  # Password hashing
python-multipart>=0.0.6  # Form data

# Monitoring & Dashboard
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
redis>=5.0.0  # Rate limiting storage

# Security
slowapi>=0.1.9  # Rate limiting
python-dotenv>=1.0.0
cryptography>=41.0.0

# Logging & SIEM
structlog>=23.2.0  # Structured logging
python-json-logger>=2.0.0

# Alerting
aiosmtplib>=3.0.0  # Email alerts
httpx>=0.25.0  # Webhook alerts

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For testing FastAPI
